Fix typos, refactor code, and enforce const correctness
Corrected minor typos in function signatures and logical statements, improving code accuracy. Refactored redundant type casts to simplify expressions and implemented const correctness for pointer parameters where applicable. Removed unused imports and cleaned up formatting issues to enhance code readability and maintainability.

+++ src/Backend/ByteCode/ByteCode.cpp
@@ -361,7 +361,7 @@ void ByteCode::makeRoomForInstructions(uint32_t room)
         return;
 
     const auto oldSize = static_cast<uint32_t>(maxInstructions * sizeof(ByteCodeInstruction));
-    maxInstructions    = std::max(static_cast<uint32_t>(numInstructions + room * 2), maxInstructions * 2);
+    maxInstructions    = std::max(numInstructions + room * 2, maxInstructions * 2);
 
     // Evaluate the first number of instructions for a given function.
     // We take the number of AST nodes in the function as a metric.
